worker/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
yt-dlp
redis
requests>=2.25
Flask>=2.0
gunicorn